superhero_project/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
pytest-django==4.5.2